refactor: modify the location of security-related file to exclude the Spring Security from the chat module
