+++ gildongmu-api/src/main/java/codeit/api/security/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package codeit.common.security;
+package codeit.api.security;
 
 import codeit.common.exception.ErrorCode;
 import codeit.domain.user.repository.UserRepository;
