+++ gildongmu-api/src/main/java/codeit/api/security/UserPrincipal.java
@@ -1,4 +1,4 @@
-package codeit.common.security;
+package codeit.api.security;
 
 import codeit.domain.user.entity.User;
 import lombok.Getter;
