+++ gildongmu-api/src/test/java/codeit/ApiApplicationTests.java
@@ -1,6 +1,5 @@
 package codeit;
 
-import codeit.api.config.SecurityConfig;
 import codeit.common.security.JwtTokenManager;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
