+++ gildongmu-api/src/test/java/codeit/api/config/WithMockOAuth2LoginUserSecurityContextFactory.java
@@ -1,6 +1,6 @@
 package codeit.api.config;
 
-import codeit.common.security.OAuth2LoginUser;
+import codeit.api.security.OAuth2LoginUser;
 import codeit.domain.user.constant.Role;
 import codeit.domain.user.entity.User;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -11,7 +11,6 @@ import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
 import org.springframework.security.test.context.support.WithSecurityContextFactory;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.Map;
 
 public class WithMockOAuth2LoginUserSecurityContextFactory implements
