+++ gildongmu-chat/src/main/java/codeit/chat/controller/ExceptionResponse.java
@@ -1,0 +1,20 @@
+package codeit.chat.controller;
+
+import lombok.Getter;
+import lombok.ToString;
+
+import java.time.LocalDateTime;
+
+@ToString
+@Getter
+public class ExceptionResponse<T> {
+    private final String code;
+    private final T message;
+    private final String timestamp;
+
+    public ExceptionResponse(String code, T message) {
+        this.code = code;
+        this.message = message;
+        this.timestamp = LocalDateTime.now().toString();
+    }
+}
